Fix NaNs appearing in gradients through HSV conversion.
When range_ or value are 0., division by zeros appear, causing NaNs.
This is not a problem when computing the function because these are masked
by jnp.where, but this causes problems during gradient computation.

Also included some test to highlight the problems before the fix. The jacobian
is only evaluated for the last image of all the TestImages, to avoid slowing
down the tests to much.
ALL_ONES and ALL_ZEROS highlight the problem.

PiperOrigin-RevId: 496848727