Add a guard to avoid NaNs appearing in the conversion from rgb to hsv.
The added test shows an encountered in the wild error, where this created NaNs
in the gradient computation. The problem seems to be related to subnormals
floating point numbers.

Annoyingly, for this to be a correct explanation, it would have been fixed by
setting the epsilon value at `jnp.finfo(image_rgb.dtype).smallest_normal`, but
for example, in the given test case, the offending numbers are ~1e-34, so these
would not be masked (`smallest_normal` being ~1e-38). Setting the threshold to
`jnp.finfo(image_rgb.dtype).eps` (~1e-7) seems to solve the issue.

PiperOrigin-RevId: 510393337